test(react-loader): fold the production gate cases into the loader suite
The new file added sanitizeResources/Ops opt-outs, which the sanitizer
baseline check rejects. The cases belong with the rest of the
SSRModuleLoader suite anyway, which already carries that opt-out for the
esbuild child process.
